+++ src/main/java/com/georgster/coinfactory/model/upgrades/tracks/automatedprecision/AiPoweredAssemblyLineUpgrade.java
@@ -13,7 +13,7 @@ public final class AiPoweredAssemblyLineUpgrade extends FactoryUpgrade {
     }
 
     public void applyUpgrade(CoinProductionState state) {
-        state.upgradeStartingProduction((long) (state.getStartingProductionValue() * 0.3));
+        state.upgradeStartingProduction(0.3);
     }
 
     /**