+++ src/main/java/com/georgster/coinfactory/model/upgrades/tracks/occultengineering/SummoningCircleOfFortuneUpgrade.java
@@ -14,14 +14,8 @@ public final class SummoningCircleOfFortuneUpgrade extends FactoryUpgrade {
     }
 
     public void applyUpgrade(CoinProductionState state) {
-        long possibleWorkingUpgrade = state.getBaseProductionValue() + 50000;
-        long possibleBaseUpgrade = state.getBaseProductionValue() * 2;
-
-        if (possibleWorkingUpgrade > possibleBaseUpgrade) { // If working upgrade is higher than base upgrade, set highest possible working increase to 75000, otherwise set it to base upgrade increase
-            state.registerHighestPossibleWorkingValue(50000);
-        } else {
-            state.registerHighestPossibleWorkingValue(state.getBaseProductionValue());
-        }
+        state.registerHighestPossibleBaseValue(50000);
+        state.registerHighestPossibleWorkingValue(1.0);
 
         int random = (int) (SoapNumbers.getRandomDouble(0, 1) * 3);
         if (random == 0) {